% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_common_theme_elements}
\alias{add_common_theme_elements}
\title{Add theme elements for visual customisation}
\usage{
add_common_theme_elements(ggplot_object)
}
\arguments{
\item{ggplot_object}{ggplot2 plot object}
}
\value{
ggplot object
}
\description{
Add theme elements for visual customisation
}
