% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{add_reference_data}
\alias{add_reference_data}
\title{Adds GEL reference data to small variant data}
\usage{
add_reference_data(small_variant_df, reference_data_list)
}
\arguments{
\item{reference_data_list}{}
}
\value{

}
\description{
Adds GEL reference data to small variant data
}
