% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{add_tmb_variant_data}
\alias{add_tmb_variant_data}
\title{Adds information from TMB trace table to small variant data}
\usage{
add_tmb_variant_data(small_variant_df, tmb_variant_df)
}
\arguments{
\item{small_variant_df}{data frame with small variants}

\item{tmb_variant_df}{data frame with TMB trace information}
}
\value{
joined data frame
}
\description{
Adds information from TMB trace table to small variant data
}
