% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{cnv}
\alias{cnv}
\title{Read in a *CopyNumberVariants.vcf file and store as an object}
\usage{
cnv(cnv_file_path, local_app = FALSE)
}
\arguments{
\item{cnv_file_path}{a file path to a *CopyNumberVariants.vcf file}
}
\value{
A cnv.output object
}
\description{
Read in a *CopyNumberVariants.vcf file
}
