% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Helper function to extract TMB/MSI metrics from
list of combined.variant.output objects}
\usage{
extract_metrics(cvo_data, category = "tmb")
}
\arguments{
\item{cvo_data}{CVO object}

\item{category}{category to extract, default: tmb}
}
\value{
data.frame
}
\description{
Helper function to extract TMB/MSI metrics from
list of combined.variant.output objects
}
