% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{filter_consequences}
\alias{filter_consequences}
\title{Filters small variant data for variant consequences. Removes
any variant with a consequence matching the list of submitted consequences}
\usage{
filter_consequences(small_variant_df, consequences)
}
\arguments{
\item{consequences}{}
}
\value{
data.frame
}
\description{
Filters small variant data for variant consequences. Removes
any variant with a consequence matching the list of submitted consequences
}
