% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{filter_for_cosmic_id}
\alias{filter_for_cosmic_id}
\title{Helper function to filter small variant data and keep
only variants with annotated COSMIC ID(s)}
\usage{
filter_for_cosmic_id(small_variant_df)
}
\arguments{
\item{small_variant_df}{}
}
\description{
Helper function to filter small variant data and keep
only variants with annotated COSMIC ID(s)
}
