% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{filter_for_Included_in_TMB}
\alias{filter_for_Included_in_TMB}
\title{Helper function to filter small variant data and keep
only variants that are included in TMB numerator}
\usage{
filter_for_Included_in_TMB(small_variant_df)
}
\arguments{
\item{small_variant_df}{data frame with small variants}
}
\value{
data frame filtered by TMB inclusion
}
\description{
Helper function to filter small variant data and keep
only variants that are included in TMB numerator
}
