% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{filter_germline_db}
\alias{filter_germline_db}
\title{Helper function to filter small variant data according
to the GermlineFilterDatabase filter}
\usage{
filter_germline_db(small_variant_df)
}
\arguments{
\item{small_variant_df}{}
}
\description{
Helper function to filter small variant data according
to the GermlineFilterDatabase filter
}
