% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{filter_germline_proxi}
\alias{filter_germline_proxi}
\title{Helper function to filter small variant data according
to the GermlineFilterProxi filter}
\usage{
filter_germline_proxi(small_variant_df)
}
\arguments{
\item{small_variant_df}{data frame with small variants}
}
\value{
data frame filtered by GermlineFilterProxi
}
\description{
Helper function to filter small variant data according
to the GermlineFilterProxi filter
}
