% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{generate_dragen_samplesheet}
\alias{generate_dragen_samplesheet}
\title{Generate sample sheet for DRAGEN TSO500 Anaylsis Pipeline input based on TSO500 sample sheet.}
\usage{
generate_dragen_samplesheet(
  samplesheet,
  file_format_version = "2",
  run_name = "RunName",
  instrument_type = "NovaSeq",
  software_version = "3.10.9",
  adapter_read1,
  adapter_read2,
  adapter_behavior = "trim",
  minimum_trimmed_read_length = 35,
  mask_short_reads = 35,
  outfile
)
}
\arguments{
\item{samplesheet}{path to TSO500 sample sheet}

\item{file_format_version}{the version of the file format (default: 2)}

\item{run_name}{the run name (default: RunName)}

\item{instrument_type}{path to output file}

\item{software_version}{The software version (default: 3.10.9)}

\item{adapter_read1}{index read1 sequence}

\item{adapter_read2}{index read2 sequence}

\item{adapter_behavior}{this indicates that the BCL Convert software trims the specified adapter sequences from each read (default: trim)}

\item{minimum_trimmed_read_length}{reads with a length trimmed below this point are masked (default: 35)}

\item{mask_short_reads}{reads with a length trimmed below this point are masked (default: 35)}

\item{outfile}{path to output CSV file}
}
\value{
the dragen analysis input sample sheet
}
\description{
Generate sample sheet for DRAGEN TSO500 Anaylsis Pipeline input based on TSO500 sample sheet.
}
