% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_analysis_details_df}
\alias{get_analysis_details_df}
\title{Extracts all analysis details from list of combined.variant.output
objects, returning a data frame of analysis details per sample}
\usage{
get_analysis_details_df(cvo_data)
}
\arguments{
\item{cvo_data}{}
}
\value{
data.frame
}
\description{
Extracts all analysis details from list of combined.variant.output
objects, returning a data frame of analysis details per sample
}
