% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{get_analysis_status.combined.quality.metrics.output}
\alias{get_analysis_status.combined.quality.metrics.output}
\title{Get analysis status from combined.quality.metrics.output object}
\usage{
\method{get_analysis_status}{combined.quality.metrics.output}(qmo_obj)
}
\arguments{
\item{qmo_obj}{qmo_obj}
}
\value{
A data frame
}
\description{
Get analysis status from combined.quality.metrics.output object
}
