% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{get_analysis_status}
\alias{get_analysis_status}
\title{Extract the analysis status from combined.quality.metrics.output object and
return in data frame format}
\usage{
get_analysis_status(qmo_obj, ...)
}
\arguments{
\item{qmo_obj}{qmo_obj}
}
\value{
A data frame with the analysis status
}
\description{
Extract the analysis status from combined.quality.metrics.output object and
return in data frame format
}
