% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_count_df}
\alias{get_count_df}
\title{Extracts the counts of the different variant types from list of combined.variant.output
objects, returning a data frame of counts per sample}
\usage{
get_count_df(cvo_data)
}
\arguments{
\item{cvo_data}{CVO object}

\item{category}{category to extract information on}
}
\value{
data frame with variant type statistics
}
\description{
Extracts the counts of the different variant types from list of combined.variant.output
objects, returning a data frame of counts per sample
}
