% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{get_dna_qc_metrics}
\alias{get_dna_qc_metrics}
\title{Extract dna qc metrics from combined.quality.metrics.output object and
return in data frame format}
\usage{
get_dna_qc_metrics(qmo_obj, ...)
}
\arguments{
\item{qmo_obj}{qmo_obj}
}
\value{
A data frame with the dna qc metrics
}
\description{
Extract dna qc metrics from combined.quality.metrics.output object and
return in data frame format
}
