% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{get_dna_qc_metrics_msi.combined.quality.metrics.output}
\alias{get_dna_qc_metrics_msi.combined.quality.metrics.output}
\title{Get dna qc metrics for msi from combined.quality.metrics.output object}
\usage{
\method{get_dna_qc_metrics_msi}{combined.quality.metrics.output}(qmo_obj)
}
\arguments{
\item{qmo_obj}{qmo_obj}
}
\value{
A data frame
}
\description{
Get dna qc metrics for msi from combined.quality.metrics.output object
}
