% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{get_dna_qc_metrics_snvtmb}
\alias{get_dna_qc_metrics_snvtmb}
\title{Extract dna qc metrics for small variants and tmb from combined.quality.metrics.output object and
return in data frame format}
\usage{
get_dna_qc_metrics_snvtmb(qmo_obj, ...)
}
\arguments{
\item{qmo_obj}{qmo_obj}
}
\value{
A data frame with dna qc metrics (small variants/TMB)
}
\description{
Extract dna qc metrics for small variants and tmb from combined.quality.metrics.output object and
return in data frame format
}
