% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{get_fusions.combined.variant.output}
\alias{get_fusions.combined.variant.output}
\title{Get fusions from combined.variant.output object}
\usage{
\method{get_fusions}{combined.variant.output}(cvo_obj)
}
\arguments{
\item{cvo_obj}{cvo_obj}
}
\value{
A data frame
}
\description{
Get fusions from combined.variant.output object
}
