% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{get_fusions}
\alias{get_fusions}
\title{Extract fusions from combined.variant.output object and
return in data frame format}
\usage{
get_fusions(cvo_obj, ...)
}
\arguments{
\item{cvo_obj}{cvo_obj}
}
\value{
A data frame of fusions
}
\description{
Extract fusions from combined.variant.output object and
return in data frame format
}
