% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{get_gene_amplifications.combined.variant.output}
\alias{get_gene_amplifications.combined.variant.output}
\title{Get gene amplifications from combined.variant.output object}
\usage{
\method{get_gene_amplifications}{combined.variant.output}(cvo_obj)
}
\arguments{
\item{cvo_obj}{cvo_obj}
}
\value{
A data frame
}
\description{
Get gene amplifications from combined.variant.output object
}
