% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{get_gene_amplifications}
\alias{get_gene_amplifications}
\title{Extract gene amplifications from combined.variant.output object and
return in data frame format}
\usage{
get_gene_amplifications(cvo_obj, ...)
}
\arguments{
\item{cvo_obj}{cvo_obj}
}
\value{
A data frame of gene amplifications
}
\description{
Extract gene amplifications from combined.variant.output object and
return in data frame format
}
