% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_metrics_df}
\alias{get_metrics_df}
\title{Extracts all TMB/MSI metrics from list of combined.variant.output
objects, returning a data frame of TMB/MSI per sample}
\usage{
get_metrics_df(cvo_data)
}
\arguments{
\item{cvo_data}{CVO object}
}
\value{
data frame with TMB/MSI metrics
}
\description{
Extracts all TMB/MSI metrics from list of combined.variant.output
objects, returning a data frame of TMB/MSI per sample
}
