% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{get_rna_expanded_metrics}
\alias{get_rna_expanded_metrics}
\title{Extract expanded rna qc metrics combined.quality.metrics.output object and
return in data frame format}
\usage{
get_rna_expanded_metrics(qmo_obj, ...)
}
\arguments{
\item{qmo_obj}{qmo_obj}
}
\value{
A data frame with expanded rna qc metrics
}
\description{
Extract expanded rna qc metrics combined.quality.metrics.output object and
return in data frame format
}
