% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{get_run_qc_metrics.combined.quality.metrics.output}
\alias{get_run_qc_metrics.combined.quality.metrics.output}
\title{Get run qc metrics from combined.quality.metrics.output object}
\usage{
\method{get_run_qc_metrics}{combined.quality.metrics.output}(qmo_obj)
}
\arguments{
\item{qmo_obj}{qmo_obj}
}
\value{
A data frame
}
\description{
Get run qc metrics from combined.quality.metrics.output object
}
