% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_sequencing_run_details_df}
\alias{get_sequencing_run_details_df}
\title{Extracts all sequencing run details from list of combined.variant.output
objects, returning a data frame of sequencing run details per sample}
\usage{
get_sequencing_run_details_df(cvo_data)
}
\arguments{
\item{cvo_data}{CVO object}
}
\value{
data frame with sequencing run details
}
\description{
Extracts all sequencing run details from list of combined.variant.output
objects, returning a data frame of sequencing run details per sample
}
