% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{get_small_variants.combined.variant.output}
\alias{get_small_variants.combined.variant.output}
\title{Get small variants from combined.variant.output object}
\usage{
\method{get_small_variants}{combined.variant.output}(cvo_obj)
}
\arguments{
\item{cvo_obj}{cvo_obj}
}
\value{
A data frame
}
\description{
Get small variants from combined.variant.output object
}
