% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{get_small_variants}
\alias{get_small_variants}
\title{Extract small variants from combined.variant.output object and
return in data frame format}
\usage{
get_small_variants(cvo_obj, ...)
}
\arguments{
\item{cvo_obj}{cvo_obj}
}
\value{
A data frame of small variants
}
\description{
Extract small variants from combined.variant.output object and
return in data frame format
}
