% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{get_splice_variants.combined.variant.output}
\alias{get_splice_variants.combined.variant.output}
\title{Get splice variants from combined.variant.output object}
\usage{
\method{get_splice_variants}{combined.variant.output}(cvo_obj)
}
\arguments{
\item{cvo_obj}{cvo_obj}
}
\value{
A data frame
}
\description{
Get splice variants from combined.variant.output object
}
