% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{get_splice_variants}
\alias{get_splice_variants}
\title{Extract splice variants from combined.variant.output object and
return in data frame format}
\usage{
get_splice_variants(cvo_obj, ...)
}
\arguments{
\item{cvo_obj}{cvo_obj}
}
\value{
A data frame of splice variants
}
\description{
Extract splice variants from combined.variant.output object and
return in data frame format
}
