% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_summarised_statistics_df}
\alias{get_summarised_statistics_df}
\title{Helper function to extract summarized counts based on sample_id}
\usage{
get_summarised_statistics_df(data_df, column_name)
}
\arguments{
\item{column_name}{}
}
\value{
data.frame
}
\description{
Helper function to extract summarized counts based on sample_id
}
