% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{get_tmb_data}
\alias{get_tmb_data}
\title{Extract TMB data from tmb.variant.output object}
\usage{
get_tmb_data(tmb_obj, ...)
}
\arguments{
\item{tmb_obj}{A tmb.variant.output object}

\item{...}{Additional arguments (not used)}
}
\value{
A data frame with TMB data
}
\description{
Extract TMB data from tmb.variant.output object
}
