% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{handle_empty_cvo_table_values}
\alias{handle_empty_cvo_table_values}
\title{Helper function to handle empty rows
in CombinedVariantOutput.tsv tabular data}
\usage{
handle_empty_cvo_table_values(intermediate_tbl)
}
\arguments{
\item{intermediate_tbl}{}
}
\value{
data.frame
}
\description{
Helper function to handle empty rows
in CombinedVariantOutput.tsv tabular data
}
