% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{keep_consequences}
\alias{keep_consequences}
\title{Filters for variant data for variant consequences. Keeps
any variant with a consequence matching the list of submitted consequences}
\usage{
keep_consequences(variant_df, consequences, type_column = "consequence_s")
}
\arguments{
\item{variant_df}{Data frame witgh variant data}

\item{consequences}{List of consequences to keep}

\item{type_column}{Name of column with variant types (default: consequence_s)}
}
\value{
data frame with consequences kept as specified
}
\description{
Filters for variant data for variant consequences. Keeps
any variant with a consequence matching the list of submitted consequences
}
