% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{make_qc_table}
\alias{make_qc_table}
\title{Visualize TSO500 QC results as gt-Table}
\usage{
make_qc_table(qc_df, id_col = "sample_id", group_name = "samples")
}
\arguments{
\item{qc_df}{data frame generated with read_xxx_qc_metrics from tso500R package}

\item{id_col}{column containing the flowcell/run_id name when multiple qc data frames were merged.}

\item{group_name}{subheader for the subtable containing the actual QC metrics}
}
\description{
Visualize TSO500 QC results as gt-Table
}
