% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{new_cnv_output}
\alias{new_cnv_output}
\title{Constructor function for combined.cnv.output objects
Not to be called directly}
\usage{
new_cnv_output(cnv_file_path, local_app = FALSE)
}
\arguments{
\item{cnv_file_path}{a file path to a *CopyNumberVariants.vcf file}

\item{local_app}{specifies whether quality metrics are coming from local app}
}
\value{
A combined.cnv.output object
}
\description{
Constructor function for combined.cnv.output objects
Not to be called directly
}
