% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{new_tmb_variant_output}
\alias{new_tmb_variant_output}
\title{Constructor function for tmb.variant.output objects
Not to be called directly}
\usage{
new_tmb_variant_output(tmb_file_path)
}
\arguments{
\item{tmb_file_path}{a file path to a TMB_trace.tsv file}
}
\value{
A tmb.variant.output object
}
\description{
Constructor function for tmb.variant.output objects
Not to be called directly
}
