% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{parse_p_dot_notation}
\alias{parse_p_dot_notation}
\title{Parses P-Dot notation column, splitting it into distinct NP ID and
amino acid variant columns}
\usage{
parse_p_dot_notation(small_variant_df)
}
\arguments{
\item{small_variant_df}{}
}
\value{

}
\description{
Parses P-Dot notation column, splitting it into distinct NP ID and
amino acid variant columns
}
