% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{parse_qmo_record}
\alias{parse_qmo_record}
\title{Helper function to parse key-value lines in MetricsOutput.tsv}
\usage{
parse_qmo_record(record_string)
}
\arguments{
\item{record_string}{}
}
\value{
char vector
}
\description{
Helper function to parse key-value lines in MetricsOutput.tsv
}
