% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{parse_qmo_table}
\alias{parse_qmo_table}
\title{Helper function to parse tabular data in MetricsOutput.tsv}
\usage{
parse_qmo_table(table_string)
}
\arguments{
\item{table_string}{}
}
\value{
data.frame
}
\description{
Helper function to parse tabular data in MetricsOutput.tsv
}
