% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{parse_record}
\alias{parse_record}
\title{Helper function to parse key-value lines in CombinedVariantOutput.tsv}
\usage{
parse_record(record_string)
}
\arguments{
\item{record_string}{}
}
\value{
char vector
}
\description{
Helper function to parse key-value lines in CombinedVariantOutput.tsv
}
