% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R, R/qualitymetrics.R
\name{parse_table}
\alias{parse_table}
\title{Helper function to parse tabular data in CombinedVariantOutput.tsv}
\usage{
parse_table(table_string)

parse_table(table_string)
}
\arguments{
\item{table_string}{}
}
\value{
data.frame

data.frame
}
\description{
Helper function to parse tabular data in CombinedVariantOutput.tsv

Helper function to parse tabular data in MetricsOutput.tsv
}
