% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_af_density}
\alias{plot_af_density}
\title{Plot allele-frequency kernel density estimate (KDE) for small variants}
\usage{
plot_af_density(small_variant_df)
}
\arguments{
\item{small_variant_df}{Data-frame containing small variants}
}
\value{
ggplot object
}
\description{
Plot allele-frequency KDE for small variants, per-sample
}
