% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_af_histogram}
\alias{plot_af_histogram}
\title{Plot allele-frequency histogram for small variants}
\usage{
plot_af_histogram(small_variant_df)
}
\arguments{
\item{small_variant_df}{Data-frame containing small variants}
}
\value{
ggplot object
}
\description{
Plot allele-frequency histogram for small variants, per-sample
}
