% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_af_per_variant_consequence}
\alias{plot_af_per_variant_consequence}
\title{Plot allele-frequency for small variants per variant consequence}
\usage{
plot_af_per_variant_consequence(small_variant_df)
}
\arguments{
\item{small_variant_df}{Data frame containing small variants}
}
\value{
ggplot object
}
\description{
Plot AF per-sample, grouped by variant consequence
}
