% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{prepare_dataframe_for_oncoprint}
\alias{prepare_dataframe_for_oncoprint}
\title{Transforms data frame holding variant information to a matrix that can be used as OncoPrint input}
\usage{
prepare_dataframe_for_oncoprint(
  variant_data_frame,
  id_column = "sample_id",
  gene_column = "gene",
  variant_type_column = "consequence_s"
)
}
\arguments{
\item{variant_data_frame}{the data frame holding variant information}
}
\value{
data.frame
}
\description{
Transforms data frame holding variant information to a matrix that can be used as OncoPrint input
}
