% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{process_small_variant_data}
\alias{process_small_variant_data}
\title{Process small variant data-frame to requirements}
\usage{
process_small_variant_data(small_variant_df)
}
\arguments{
\item{small_variant_df}{Data-frame of small variants}
}
\value{

}
\description{
Processes small-variant data to comply with requirements for
further analysis. The function:

    1. filters for variants that:
      - have a consequence in a pre-defined list (see details)
      - are present with depth > 0
    2. extracts NP ID and protein information from the P Dot-notation column
    3. adds columns to faciliate addition of reference data
}
\details{
The following variant consequences are currently included:
- frameshift_variant
- inframe_deletion
- inframe_insertion
- missense_variant
- missense_variant:splice_region_variant
- splice_acceptor_variant
- splice_donor_variant
- splice_donor_variant:intron_variant
- start_lost
- stop_gained
- stop_gained:splice_region_variant
- stop_lost
}
