% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{read_analysis_status}
\alias{read_analysis_status}
\title{Read in a list of combined.quality.metrics.output objects
and return a data frame of the analysis status per timepoint}
\usage{
read_analysis_status(qmo_list)
}
\arguments{
\item{qmo_list}{a list of combined.quality.metrics.output objects}
}
\value{
A data frame of analysis status extracted from
each combined.quality.metrics.output object, per timepoint
}
\description{
Read in a list of combined.quality.metrics.output objects
and return a data frame of the analysis status per timepoint
}
