% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{read_annotation_data}
\alias{read_annotation_data}
\title{Read GEL annotation Excel spreadsheet into a list of data frames, per sheet}
\usage{
read_annotation_data(annotation_data_path)
}
\arguments{
\item{annotation_data_path}{Path to annotation .xlsx}
}
\value{
list of data.frame
}
\description{
Read GEL annotation Excel spreadsheet into a list of data frames, per sheet
}
