% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{read_cnv_data}
\alias{read_cnv_data}
\title{Read in a batch of *CopyNumberVariants.vcf files into a list of CNV objects}
\usage{
read_cnv_data(cnv_directory, local_app = FALSE)
}
\arguments{
\item{cnv_directory}{a file path to a directory containing one of more *CopyNumberVariants.vcf files}

\item{local_app}{specifies whether quality metrics are coming from local app}
}
\value{
A named list of combined.cnv.output objects
}
\description{
Read in a batch of *CopyNumberVariants.vcf files into a list of CNV objects
}
