% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{read_cvo_data}
\alias{read_cvo_data}
\title{Read in a batch of CombinedVariantOutput.tsv files into a list}
\usage{
read_cvo_data(cvo_directory, local_app = FALSE)
}
\arguments{
\item{cvo_directory}{a file path to a directory containing one of more CombinedVariantOutput.tsv files}

\item{local_app}{specifies whether quality metrics are coming from local app}
}
\value{
A named list of combined.variant.output objects
}
\description{
Read in a batch of CombinedVariantOutput.tsv files into a list
}
