% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{read_dna_expanded_metrics}
\alias{read_dna_expanded_metrics}
\title{Read in a list of combined.quality.metrics.output objects
and return a data frame of expanded dna qc metrics per timepoint}
\usage{
read_dna_expanded_metrics(qmo_list)
}
\arguments{
\item{cvo_list}{a list of combined.quality.metrics.output objects}
}
\value{
A data frame of expanded dna qc metrics extracted from
each combined.quality.metrics.output object, per timepoint
}
\description{
Read in a list of combined.quality.metrics.output objects
and return a data frame of expanded dna qc metrics per timepoint
}
