% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{read_fusions}
\alias{read_fusions}
\title{Read in a list of combined.variant.output objects
and return a data frame of fusions per sample}
\usage{
read_fusions(cvo_list)
}
\arguments{
\item{cvo_list}{a list of combined.variant.output objects}
}
\value{
A data frame of fusions extracted from
each combined.variant.output object, per sample
}
\description{
Read in a list of combined.variant.output objects
and return a data frame of fusions per sample
}
