% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{read_gene_amplifications}
\alias{read_gene_amplifications}
\title{Read in a list of combined.variant.output objects
and return a data frame of gene amplifications per sample}
\usage{
read_gene_amplifications(cvo_list)
}
\arguments{
\item{cvo_list}{a list of combined.variant.output objects}
}
\value{
A data frame of gene amplifications extracted from
each combined.variant.output object, per sample
}
\description{
Read in a list of combined.variant.output objects
and return a data frame of gene amplifications per sample
}
