% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{read_qmo_data}
\alias{read_qmo_data}
\title{Read in a batch of MetricsOutput.tsv files into a list}
\usage{
read_qmo_data(qmo_directory, local_app = FALSE)
}
\arguments{
\item{qmo_directory}{a file path to a directory containing one of more
MetricsOutput.tsv files}

\item{local_app}{specifies whether quality metrics are coming from local app (default: FALSE)}
}
\value{
A named list of combined.quality.metrics.output objects
}
\description{
Read in a batch of MetricsOutput.tsv files into a list
}
