% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{read_reference}
\alias{read_reference}
\title{Read GEL reference Excel spreadsheet into a list of data frames, per sheet}
\usage{
read_reference(reference_path)
}
\arguments{
\item{reference_path}{Path to reference .xlsx}
}
\value{
list of data.frame
}
\description{
Read GEL reference Excel spreadsheet into a list of data frames, per sheet
}
