% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{read_run_qc_metrics}
\alias{read_run_qc_metrics}
\title{Read in a list of combined.quality.metrics.output objects
and return a data frame of run qc metrics per timepoint}
\usage{
read_run_qc_metrics(qmo_list)
}
\arguments{
\item{qmo_list}{a list of combined.quality.metrics.output objects}
}
\value{
A data frame of small variants extracted from
each combined.quality.metrics.output object, per timepoint
}
\description{
Read in a list of combined.quality.metrics.output objects
and return a data frame of run qc metrics per timepoint
}
