% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{read_splice_variants}
\alias{read_splice_variants}
\title{Read in a list of combined.variant.output objects
and return a data frame of splice variants per sample}
\usage{
read_splice_variants(cvo_list)
}
\arguments{
\item{cvo_list}{a list of combined.variant.output objects}
}
\value{
A data frame of splice variants extracted from
each combined.variant.output object, per sample
}
\description{
Read in a list of combined.variant.output objects
and return a data frame of splice variants per sample
}
