% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{read_tmb_trace_data}
\alias{read_tmb_trace_data}
\title{Read in a batch of TMB_trace.tsv files into a list}
\usage{
read_tmb_trace_data(tmb_directory)
}
\arguments{
\item{tmb_directory}{a file path to a directory containing one of more
TMB_trace.tsv files}
}
\value{
A named list of data frame objects
}
\description{
Read in a batch of TMB_trace.tsv files into a list
}
