% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{summarize_cnv_data}
\alias{summarize_cnv_data}
\title{Read in a batch of *CopyNumberVariants.vcf files into one dataframe}
\usage{
summarize_cnv_data(cnv_directory)
}
\arguments{
\item{tmb_directory}{a file path to a directory containing one of more
*tmb.json files}
}
\value{
A dataframe with the read CNV data
}
\description{
Read in a batch of *CopyNumberVariants.vcf files into one dataframe
}
