% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R, R/tmb_optimized.R
\name{tmb}
\alias{tmb}
\title{Read in a TMB_trace.tsv file and store as an object}
\usage{
tmb(tmb_file_path)

tmb(tmb_file_path)
}
\arguments{
\item{tmb_file_path}{a file path to a TMB_trace.tsv file}
}
\value{
A tmb.variant.output object

A tmb.variant.output object
}
\description{
Read in a TMB_trace.tsv file

Read in a TMB_trace.tsv file
}
\examples{
\dontrun{
# Read a single TMB trace file
tmb_obj <- tmb("/path/to/TMB_trace.tsv")
print(tmb_obj)

# Extract data from TMB object
tmb_data <- get_tmb_data(tmb_obj)

# Read multiple TMB trace files
tmb_data <- read_tmb_trace_data("/path/to/tmb/directory")

# Summarize TMB data
summary_stats <- summarize_tmb_data(tmb_data)

# Filter TMB data
filtered_data <- filter_tmb_data(tmb_data, min_tmb = 5, max_tmb = 50)
}

}
\seealso{
\link{cvo}, \link{cnv}, \link{qualitymetrics}
}
