% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R, R/qualitymetrics.R
\name{trim_header_and_footer}
\alias{trim_header_and_footer}
\title{Helper function to remove header and footer from CombinedVariantOutput.tsv}
\usage{
trim_header_and_footer(string)

trim_header_and_footer(string)
}
\arguments{
\item{string}{}
}
\value{
char vector

char vector
}
\description{
Helper function to remove header and footer from CombinedVariantOutput.tsv

Helper function to remove header and footer from MetricsOutput.tsv
}
