% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{trim_qmo_header_and_footer}
\alias{trim_qmo_header_and_footer}
\title{Helper function to remove header and footer from MetricsOutput.tsv}
\usage{
trim_qmo_header_and_footer(string)
}
\arguments{
\item{string}{}
}
\value{
char vector
}
\description{
Helper function to remove header and footer from MetricsOutput.tsv
}
