% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{update_annotation_join_columns}
\alias{update_annotation_join_columns}
\title{Helper function to make small variant DF joinable to annotation data}
\usage{
update_annotation_join_columns(small_variant_df)
}
\arguments{
\item{small_variant_df}{data frame with small variants}
}
\value{
data frame
}
\description{
Helper function to make small variant DF joinable to annotation data
}
