% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{update_ref_join_columns}
\alias{update_ref_join_columns}
\title{Helper function to make small variant DF joinable to reference data}
\usage{
update_ref_join_columns(small_variant_df)
}
\arguments{
\item{small_variant_df}{}
}
\value{
data frame
}
\description{
Helper function to make small variant DF joinable to reference data
}
