% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{validate_tso500}
\alias{validate_tso500}
\title{Validator function for combined.variant.output constructor
Not to be called directly
NOT IMPLEMENTED}
\usage{
validate_tso500()
}
\value{
boolean
}
\description{
Validator function for combined.variant.output constructor
Not to be called directly
NOT IMPLEMENTED
}
