% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{validate_tso500_qc}
\alias{validate_tso500_qc}
\title{Validator function for quality.metrics.output constructor
Not to be called directly
NOT IMPLEMENTED}
\usage{
validate_tso500_qc()
}
\value{
boolean
}
\description{
Validator function for quality.metrics.output constructor
Not to be called directly
NOT IMPLEMENTED
}
