% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_multiqc_data}
\alias{write_multiqc_data}
\title{Write files needed for reporting in MultiQC.}
\usage{
write_multiqc_data(
  analysis_details_frame,
  sequencing_run_details_frame,
  variant_stats,
  tmb_msi_data_frame,
  folder_path
)
}
\arguments{
\item{analysis_details_frame}{data frame with analysis details (get_analysis_details_df)}

\item{sequencing_run_details_frame}{data frame with sequencing run details (get_sequencing_run_details_df)}

\item{variant_stats}{data frame with variant type statistics (get_count_df)}

\item{tmb_msi_data_frame}{data frame with tmb and msi statistics (get_metrics_df)}

\item{file_path}{path to output file}
}
\value{
the files needed for generating a MultiQC report
}
\description{
Write files needed for reporting in MultiQC.
}
