% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_rdata_file}
\alias{write_rdata_file}
\title{Save data frames to RData object.}
\usage{
write_rdata_file(data_frames, file_path)
}
\arguments{
\item{data_frames}{list of data frames}

\item{file_path}{path to output file}
}
\description{
Save data frames to RData object.
}
