% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_workbook}
\alias{write_workbook}
\title{Save data frames to Microsoft Excel workbook}
\usage{
write_workbook(workbook_name, data_frames, sheet_names)
}
\arguments{
\item{workbook_name}{Filename of the workbook}

\item{data_frames}{list of data frames}

\item{sheet_names}{character vector of worksheet names}
}
\value{
the Excel workbook
}
\description{
Store data frames to Microsoft Excel workbook. Submission of
  more than one data frame will result in a workbook with that many sheets.
  The number of worksheet names _must_ match the number of data frames
  submitted.
}
