% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_worksheet}
\alias{write_worksheet}
\title{Write worksheet to Excel workbook.}
\usage{
write_worksheet(data, sheet_name, sheet_index, workbook)
}
\arguments{
\item{workbook}{}
}
\value{

}
\description{
Write worksheet to Excel workbook.
}
