% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{add_amplification_data}
\alias{add_amplification_data}
\title{Adds amplifications to small variant data, 
variant type (consequence_s) column renamed to variant_type}
\usage{
add_amplification_data(small_variant_df, amplification_df)
}
\arguments{
\item{small_variant_df}{data frame with small variants}

\item{amplification_df}{data frame with amplifications}
}
\value{
joined data frame
}
\description{
Adds amplifications to small variant data, 
variant type (consequence_s) column renamed to variant_type
}
