% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{add_annotation_data}
\alias{add_annotation_data}
\title{Adds GEL annotation data to small variant data}
\usage{
add_annotation_data(small_variant_df, annotation_data_list)
}
\arguments{
\item{annotation_data_list}{}
}
\description{
Adds GEL annotation data to small variant data
}
