% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{cvo}
\alias{cvo}
\title{Read in a CombinedVariantOutput.tsv file and store as an object}
\usage{
cvo(cvo_file_path, local_app = FALSE, ctdna = FALSE)
}
\arguments{
\item{cvo_file_path}{a file path to a CombinedVariantOutput.tsv file}

\item{local_app}{specifies whether data has been generated using the local app}

\item{ctdna}{specifies whether data has been generated using ctdna workflow}
}
\value{
A combined.variant.output object
}
\description{
Read in a CombinedVariantOutput.tsv file
}
