% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{filter_depth}
\alias{filter_depth}
\title{Helper function to filter small variant data according
to specified depth}
\usage{
filter_depth(small_variant_df, depth_limit = 0)
}
\arguments{
\item{depth_limit}{}
}
\value{

}
\description{
Helper function to filter small variant data according
to specified depth
}
