% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{new_combined_quality_metrics_output}
\alias{new_combined_quality_metrics_output}
\title{Constructor function for quality.metrics.output objects
Not to be called directly}
\usage{
new_combined_quality_metrics_output(metrics_file_path, local_app = FALSE)
}
\arguments{
\item{metrics_file_path}{a file path to a MetricsOutput.tsv file}

\item{local_app}{specifies whether quality metrics are coming from local app}
}
\value{
A quality.metrics.output object
}
\description{
Constructor function for quality.metrics.output objects
Not to be called directly
}
