% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvo.R
\name{new_combined_variant_output}
\alias{new_combined_variant_output}
\title{Constructor function for combined.variant.output objects
Not to be called directly}
\usage{
new_combined_variant_output(cvo_file_path)
}
\arguments{
\item{cvo_file_path}{a file path to a CombinedVariantOutput.tsv file}
}
\value{
A combined.variant.output object
}
\description{
Constructor function for combined.variant.output objects
Not to be called directly
}
