% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{parse_illumina_samplesheet}
\alias{parse_illumina_samplesheet}
\title{Parses DRAGEN sample sheet and returns a dataframe}
\usage{
parse_illumina_samplesheet(samplesheet)
}
\arguments{
\item{samplesheet}{path to Illumina sample sheet}

\item{id_column}{Column holding identifiers}

\item{gene_column}{Column holding genes}

\item{variant_type_column}{Column holding variant types}
}
\value{
names list of data frames (header, settings, data) holding sample sheet content
}
\description{
Parses DRAGEN sample sheet and returns a dataframe
}
