% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_onco_print}
\alias{plot_onco_print}
\title{Plot OncoPrint plot (heatmap) for variants}
\usage{
plot_onco_print(
  variant_matrix,
  column_title,
  alter_list,
  variant_colors,
  heatmap_legend,
  bottom_annotation = anno_empty(border = FALSE),
  top_annotation = anno_empty(border = FALSE),
  right_annotation = anno_empty(border = FALSE),
  left_annotation = anno_empty(border = FALSE),
  top = nrow(variant_matrix)
)
}
\arguments{
\item{column_title}{Title of OncoPrint plot}

\item{alter_list}{List with alter_graphics functions for variant types}

\item{variant_colors}{List of defined colors for variant types}

\item{heatmap_legend}{List containing legend specs (title, at, labels)}

\item{top}{Plot top {top} genes for given variant matrix (default: all)}

\item{small_variant_df}{Data frame containing variants for multiple samples}
}
\value{
ggplot object
}
\description{
Plot OncoPrint plot for specified list of variants
}
