% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{prepare_dataframe_for_oncoprint}
\alias{prepare_dataframe_for_oncoprint}
\title{Transforms data frame holding variant information to a matrix that can be used as OncoPrint input}
\usage{
prepare_dataframe_for_oncoprint(
  variant_data_frame,
  id_column = "sample_id",
  gene_column = "gene",
  variant_type_column = "consequence_s"
)
}
\arguments{
\item{variant_data_frame}{Data frame holding variant information}

\item{id_column}{Column holding identifiers}

\item{gene_column}{Column holding genes}

\item{variant_type_column}{Column holding variant types}
}
\value{
data frame as needed for onocprint plot
}
\description{
Transforms data frame holding variant information to a matrix that can be used as OncoPrint input
}
