% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitymetrics.R
\name{qualitymetrics}
\alias{qualitymetrics}
\title{Read in a MetricsOutput.tsv file and store as an object}
\usage{
qualitymetrics(metrics_file_path)
}
\arguments{
\item{metrics_file_path}{a file path to a MetricsOutput.tsv file}
}
\value{
A quality.metrics.output object
}
\description{
Read in a MetricsOutput.tsv file
}
