% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{read_annotation_data}
\alias{read_annotation_data}
\title{Read Excel spreadsheet into a list of data frames, per sheet}
\usage{
read_annotation_data(annotation_data_path, sheet_names)
}
\arguments{
\item{annotation_data_path}{Path to annotation .xlsx}

\item{sheet_names}{List of names of the sheets in the Excel spreadsheet}
}
\value{
list of data.frame
}
\description{
Read Excel spreadsheet into a list of data frames, per sheet
}
