% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{read_tmb_details_data}
\alias{read_tmb_details_data}
\title{Read in a batch of *tmb.json files into a list}
\usage{
read_tmb_details_data(tmb_directory)
}
\arguments{
\item{tmb_directory}{a file path to a directory containing one of more
*tmb.json files}
}
\value{
A named list of data frame objects
}
\description{
Read in a batch of *tmb.json files into a list
}
