% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{tmb}
\alias{tmb}
\title{Read in a TMB_trace.tsv file and store as an object}
\usage{
tmb(tmb_file_path)
}
\arguments{
\item{tmb_file_path}{a file path to a TMB_trace.tsv file}
}
\value{
A tmb.variant.output object
}
\description{
Read in a TMB_trace.tsv file
}
